Replace uint8 with uint256

+++ contracts/GaltToken.sol
@@ -24,7 +24,7 @@ contract GaltToken is ERC20Mintable {
   string public constant _symbol = "GALT";
 
   // solium-disable-next-line uppercase
-  uint8 public constant _decimals = 18;
+  uint256 public constant _decimals = 18;
 
   uint256 public constant INITIAL_SUPPLY = 0;
